Added support for an external script editor, specified in Preferences (currently at the bottom of the General section).  It can be invoked by the CLI "edit" command, or by selecting an item from the Scripts menu with the shift key held down.
Note that, on the Mac at least, the shift key has to be down when the menu is
first pulled down - pressing it while the menu is already open doesn't work,
and I see no way to fix that.

This has only been tested with TextWrangler, in which case the editor command
is simply 'edit' (assuming that you let it install its command-line tools).


git-svn-id: https://svn.code.sf.net/p/kolmafia/code@8079 b29ace70-8910-0410-8dcc-aa2fc6433167