Add a "volcano solve" command. This visits the volcano cave to figure out where you are and will take up to four random moves to find all the platform postions. Then, it performs a breadth-first search of the puzzle to figure out how to get to the goal from wherever it ended up after the initial exploration. It's possible that the random moves left you in an impossible spot. If so, it detects that and tells you to swim to shore and try again. If you do so, there WILL be a solution.
For now, this command simply prints the list of hops you need to take to get to
the goal from wherever it left you after the initial exploration.

Soon, it will actually execute the moves and leave you one hop away from the
oal.

Note that there is no knowledge whatsoever of the "standard" configurations.
Nor is there need for you to be at the starting square. This algorithm is
perfectly capable of adapting to new configurations and can handle you having
moved to arbitrary places before calling upon KoLmafia to finish for you.


git-svn-id: https://svn.code.sf.net/p/kolmafia/code@8215 b29ace70-8910-0410-8dcc-aa2fc6433167