Preliminary support for the revamped Greater-Than Sign: new items, minimal monster data, choice adventure, and removal of the unidentified version of items from your inventory after reading the plus sign (that part hasn't been tested yet).
I've set the default conditions for this area to the slightly unusual "1 plus
sign, +1 steaming evil".  The idea is that you would leave the choice
adventure set to the option that gives you Teleportitis.  Adventuring here
will get the plus sign on the first choiceadv (because the goal overrides the
setting), then get Teleportitis on the second, and continue until the Oracle
is found - which is a stop adventure.  Putting an unobtainable goal in the
conditions seemed the easiest way to accomplish this.


git-svn-id: https://svn.code.sf.net/p/kolmafia/code@8296 b29ace70-8910-0410-8dcc-aa2fc6433167