Some changes to make it easier to write scripts that present an actual user interface, in the relay browser:
* Any .ash scripts in the relay folder, with names starting with "relay_",
are collected in a menu shown at the bottom right of the top pane.  The list
is rebuilt whenever the top pane is loaded, so the [re] button that mafia
already adds there works to refresh the scripts, too. Selecting one loads it
in the main pane.  The relay browser override feature does NOT have to be
enabled for this to work.

* Added ASH predefined string constant __FILE__, which is the filename of the
current script.  This allows relay scripts to link or submit a form to
themselves, even if the user has renamed them.

* Added ASH function form_fields(), which returns a string[string] map with
all the name/value pairs from the relay request being handled.  Unlike the
existing form_field() function, this allows iterating over all the fields,
and can distinguish a field with a blank value from a nonexistent field.

* Added ASH string functions entity_encode(string) and entity_decode(string),
for converting text that might contain special characters such as ampersands
and angle brackets so that it can safely be displayed in HTML, and for
retrieving such text from form fields.


git-svn-id: https://svn.code.sf.net/p/kolmafia/code@8318 b29ace70-8910-0410-8dcc-aa2fc6433167