More ConsequenceManager stuff: Added patterns to update mafia's idea of the current monster that is trailed, puttied, or camera-ed when the corresponding description is viewed.
Added capability to apply a consequence to the initial page load of a combat,
possibly returning a more specific monster name.  Lots of previously
hard-coded monster disambiguations are now in this format, and a new one is
added: Ed now has all 7 undying forms disambiguated, with correct HP values
for each.


git-svn-id: https://svn.code.sf.net/p/kolmafia/code@8372 b29ace70-8910-0410-8dcc-aa2fc6433167