Rather than have GenericRequest and KoLAdventure extend Job, have them extend Runnable instead.
This means that every "request" in KoLmafia will be run inside of a runnable wrapper.  In this way, we can put everything inside of a try-catch and see if there's any exceptions that are happening if a request is run in a worker thread that might not be getting reported in a debug log.

git-svn-id: https://svn.code.sf.net/p/kolmafia/code@9719 b29ace70-8910-0410-8dcc-aa2fc6433167