Maximizer tweaks: * +/-melee now actually requires a weapon of the indicated type, rather than merely disallowing a weapon of the other type (an inadvertent change which crept in with the sudden popularity of unarmed combat).  If you really want the previous behavior, give a weight of +/-0.5.
* 15% of your weapon's power (including dual-wielded and Hand weapons) is now
added to your Weapon Damage modifier, approximating the extra damage you get
based on the power.  This gives the Maximizer a reason to prefer
higher-powered weapons.

* All of the oddities related to wielding a chefstaff as a Sauceror without
Rigatoni should now be fixed, which mainly involved realizing that the
special sauce glove is an accessory, not an offhand.  In particular, the
glove will be equipped first if it's needed.

* Maximizer keywords can now be enclosed in double quotes, allowing them to
contain numbers and commas.  Note that the weight isn't quoted.  For example:
    +"equip 17-alarm saucepan"


git-svn-id: https://svn.code.sf.net/p/kolmafia/code@10531 b29ace70-8910-0410-8dcc-aa2fc6433167