Implement unsticky scroll panes using a non-static variable, since calling dispose() before resulted in all known references clearing and a potential null pointer exception.
Use a linked list instead of a hash map of boolean values for sticky scroll panes to avoid checking a map all the time.

Probably over-defensive, but retain the pattern of using weak references and wrap sticky panes with a weak reference in case of a Swing memory leak that might happen if dispose() is not properly called.

git-svn-id: https://svn.code.sf.net/p/kolmafia/code@10848 b29ace70-8910-0410-8dcc-aa2fc6433167